Fix gcc error when compiling link scripts.
At some point between gcc11 and 14 gcc now requires by default that main() is
declared to return type int.  This likely would have warned before but now
causes an error if the int is missing.  The test program in the configure
cript did not have the int declaration, this commit fixes that and should allow
compile link to work across all gcc versions to date.